+++ persistent-postgresql/ChangeLog.md
@@ -1,5 +1,10 @@
 # Changelog for persistent-postgresql
 
+## 2.13.6.0
+
+* [#xxx](FILL_THIS_IN_LATER)
+   * Add `isSerializationFailure` and `isDeadlockDetected` exception predicates
+
 ## 2.13.5.2
 
 * [#1471](https://github.com/yesodweb/persistent/pull/1471)